Added passport auth logic, code rafactoring #53
